Implement numeric casts, improve optimization diagnostics, and make a little progress in toasm

+++ trunk/abc/cond.c
@@ -1,6 +1,6 @@
 
 int main(int argc, char **argv) {
-  if (argc > 5) {
+  if (argc > (short)5) {
     return 1;
   } else {
     return 2;
